backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "whatif-backend"
version = "0.1.0"
description = "AI Film Studio backend and ingestion pipeline"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.20.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "jouvence>=0.4.0",
    "google-genai>=1.0.0",
    "pymupdf>=1.24.0",
]

[project.scripts]
whatif = "ingestion.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["ingestion"]
